requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
autodistill==0.1.28
blinker==1.8.2
certifi==2023.7.22
cffi==1.17.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.10.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
easyocr==1.7.1
executing==2.0.1
fastapi==0.112.1
ffmpy==0.4.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gradio==4.41.0
gradio_client==1.3.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==2.10
imageio==2.35.1
importlib_resources==6.4.3
influxdb-client==1.45.0
ipykernel
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lapx==0.5.9.post1
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.1
reactivex==4.0.4
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.27
ruff==0.6.1
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
setuptools==73.0.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.2
supervision==0.22.0
sympy==1.13.2
tifffile==2024.8.10
tomlkit==0.12.0
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.79
ultralytics-thop==2.0.5
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
yarl==1.11.0
